[Kernel_Arguments] Add parameter in get_arguments to identify the unit type
How the arguments are passed now depends on the unit type. I left for now
the "x" name for PEAK (we have to ask to UPV for this). If the kernel
is a GN, we pass the kernel binary path and the size of GN memory,
followed by the addresses of buffers, as offset from file beginning.

Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>